Update issue 182 Labels: -Milestone-Release1.8 +Milestone-Release1.9 If only the state is changed, then the revision number will not increase unless the checkbox next to the save button is checked.
New issue
Summary: Only increase revision number on save if approved by user
At the moment all edit to an entity results in a new revision. This should be
changed, so new revisions only occure if the user specifies it. This involves
quite a big change in the code, since almost all code just inserts new entries
in the DB instead of updating existing.